finish eru-agent install and modify

+++ armin/service.py
@@ -6,7 +6,10 @@ import logging
 from armin import config as armin_config
 from armin.services import docker
 from armin.services import mfsmount
+from armin.services import eru_agent
 from armin.services import falcon_agent
+
+from armin.services import get_service
 
 logger = logging.getLogger(__name__)
 
@@ -26,10 +29,11 @@ class Service(object):
             cmd = 'systemctl stop mfsmount && yum update moosefs-client -y && systemctl start mfsmount'
             return self.host._execute(cmd)
         if config:
-            svr = mfsmount.MFSmount(config)
+            svr = get_service(mfsmount.MFSmount, config)
             if not svr:
                 return False
-            svr.make_service(self.host._upload)
+            if not svr.make_service(self.host._upload):
+                return False
             cmd = 'systemctl stop mfsmount'
             if update:
                 cmd += ' && yum update moosefs-client -y'
@@ -43,10 +47,11 @@ class Service(object):
             cmd = 'systemctl stop falcon-agent && yum update falcon-agent -y && systemctl start falcon-agent'
             return self.host._execute(cmd)
         if config:
-            svr = falcon_agent.FalconAgent(config)
+            svr = get_service(falcon_agent.FalconAgent, config)
             if not svr:
                 return False
-            svr.make_service(self.host._upload)
+            if not svr.make_service(self.host._upload):
+                return False
             cmd = 'systemctl stop falcon-agent'
             if update:
                 cmd += ' && yum update falcon-agent -y'
@@ -61,21 +66,43 @@ class Service(object):
         cmd = 'yum update docker-engine -y && systemctl restart docker'
         return self.host._execute(cmd)
 
+    def modify_eru_agent(self, update=False, config=None):
+        if update and not config:
+            cmd = 'systemctl stop eru-agent && yum update eru-agent -y && systemctl start eru-agent'
+            return self.host._execute(cmd)
+        if config:
+            svr = get_service(eru_agent.EruAgent, config)
+            if not svr:
+                return False
+            if not svr.make_service(self.host._upload):
+                return False
+            cmd = 'systemctl stop eru-agent'
+            if update:
+                cmd += ' && yum update eru-agent -y'
+            cmd += ' && systemctl start eru-agent'
+            return self.host._execute(cmd)
+        return False
+
     def install_moosefs_client(self, config, enable=False):
-        svr = mfsmount.MFSmount(config)
+        svr = get_service(mfsmount.MFSmount, config)
         if not svr:
             return False
-        svr.make_service(self.host._upload)
+        if not svr.make_service(self.host._upload):
+            return False
         cmd = 'yum install moosefs-client -y && mkdir -p %s && systemctl daemon-reload && systemctl start mfsmount' % svr.mount
         if enable:
             cmd += ' && systemctl enable mfsmount'
         return self.host._execute(cmd)
 
     def install_falcon_agent(self, config, enable=False):
-        svr = falcon_agent.FalconAgent(config)
+        svr = get_service(falcon_agent.FalconAgent, config)
         if not svr:
             return False
-        svr.make_service(self.host._upload)
+        cmd = 'mkdir -p %s' % armin_config.REMOTE_FALCON_AGENT_CONFIG_PATH
+        if not self.host._execute(cmd):
+            return False
+        if not svr.make_service(self.host._upload):
+            return False
         cmd = 'yum install falcon-agent -y && systemctl start falcon-agent'
         if enable:
             cmd += ' && systemctl enable falcon-agent'
@@ -83,7 +110,7 @@ class Service(object):
 
     def install_docker(self, config, enable=False):
         config['ip'] = self.host.server
-        svr = docker.Docker(config)
+        svr = get_service(docker.Docker, config)
         if not svr:
             return False
         if not svr.make_service(self.host._upload):
@@ -104,5 +131,19 @@ class Service(object):
         )
         if enable:
             cmd += ' && systemctl enable docker'
+        return self.host._execute(cmd)
+
+    def install_eru_agent(self, config, enable=False):
+        svr = get_service(eru_agent.EruAgent, config)
+        if not svr:
+            return False
+        cmd = 'mkdir -p %s' % armin_config.REMOTE_ERU_AGENT_CONFIG_PATH
+        if not self.host._execute(cmd):
+            return False
+        if not svr.make_service(self.host._upload):
+            return False
+        cmd = 'yum install eru-agent -y && systemctl start eru-agent'
+        if enable:
+            cmd += ' && systemctl enable eru-agent'
         return self.host._execute(cmd)
